Compiles but does not work: inferences are commented out
Added makefile support for setting the name of the provided libraries (currently idp_intern) instead of them scattered through the code;

Refactoring execute such that
	adding inferences is easier
	as much code as possible is really independent of the lua interface
	code reduction by using templates and refactoring out methods
	created hard mapping between argtypes and their names

+++ ground.hpp
@@ -261,7 +261,7 @@ class GroundTranslator {
 		int				nextNumber();
 		unsigned int	addSymbol(PFSymbol* pfs);
 
-		bool										isSymbol(int nr)			const	{ return 0<nr && nr<_backsymbtable.size(); }
+		bool					isSymbol(int nr)			const	{ return 0<nr && (uint)nr<_backsymbtable.size(); }
 		PFSymbol*				symbol(int nr)				const	{ return _backsymbtable[abs(nr)];			}
 		const ElementTuple&		args(int nr)				const	{ return _backargstable[abs(nr)];			}
 		bool					isTseitin(int l)			const	{ return symbol(l) == 0;					}
